apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.18
httpx==0.28.0
redis[hiredis]==5.2.0
celery[redis]==5.4.0
structlog==24.4.0
stripe>=8.0.0
cryptography==44.0.0
PyJWT[crypto]==2.10.1
aioboto3==13.2.0
python-dotenv==1.0.1
email-validator==2.2.0
langgraph==0.2.56
langchain-anthropic==0.3.0
langchain-core==0.3.25
openai>=1.40.0
